pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "free-transformer"
version = "0.1.0"
description = "A modular implementation of the Free Transformer architecture with conditional VAE"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Uday Phalak", email = "iam.udayphalak@gmail.com"}
]
keywords = ["transformer", "vae", "deep-learning", "pytorch", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "datasets>=2.14.0",
    "transformers>=4.30.0",
    "sentencepiece>=0.1.99",
    "tiktoken>=0.5.0",
    "tensorboard>=2.13.0",
    "wandb>=0.15.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "mypy>=1.4.0",
    "flake8>=6.0.0",
    "ruff>=0.14.3",
    "types-pyyaml>=6.0.12.20250915",
    "isort>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "mypy>=1.4.0",
    "flake8>=6.0.0",
    "ipython>=8.14.0",
]
deepspeed = [
    "deepspeed>=0.10.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
    "mkdocstrings[python]>=0.24.0",
    "mkdocs-autorefs>=0.5.0",
]

[project.urls]
Homepage = "https://github.com/udapy/free-transformer"
Documentation = "https://free-transformer.readthedocs.io"
Repository = "https://github.com/udapy/free-transformer"

[project.scripts]
free-transformer = "free_transformer.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=src/free_transformer --cov-report=html --cov-report=term"
